# openapi.yaml
openapi: 3.0.3
info:
  title: CI/CD Platform API
  description: |
    API REST du backend Laravel pour la plateforme CI/CD.

    **Authentification** : Token Bearer via `Authorization: Bearer <token>`.
    Obtenir le token via `POST /api/login`.

    **Rôles** :
    - `administrateur` — gestion complète
    - `administrateur_cloud_doi` — déploiements CD
    - `securite` — lecture seule
  version: "1.0.0"
servers:
  - url: http://10.100.224.70:8000/api
    description: Serveur de développement
  - url: http://localhost:8000/api
    description: Local
security:
  - BearerAuth: []
...
paths:
...
  /deploiements/{id}/logs:
    get:
      tags: [Déploiements]
      summary: Détail complet d'un déploiement avec logs compilés CI et CD
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer}}
      responses:
        '200':
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DeploiementDetail'}

  # ───── LOGS ──────────────────────────────────────────────
...
  /projets/{id}/pipelines/{runId}/logs:
    get:
      tags: [Logs]
      summary: Logs CI + CD fusionnés pour un run GitHub Actions
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer}}
        - {name: runId, in: path, required: true, schema: {type: integer}}
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  run_id:            {type: integer}
                  sections:
                    type: array
                    items:
                      properties:
                        source: {type: string, enum: [CI,CD]}
                        job:    {type: string}
                        lignes:
                          type: array
                          items:
                            properties:
                              timestamp: {type: string, nullable: true}
                              texte:     {type: string}
                  cd_statut:          {type: string, nullable: true}
                  cd_deploiement_id:  {type: integer, nullable: true}
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Token SHA-256 obtenu via POST /api/login
  schemas:
...
    DeploiementListItem:
      type: object
      properties:
        id:                      {type: integer}
        deployment_id:           {type: string, nullable: true}
        ci_run_id:               {type: integer, nullable: true}
        cd_run_id:               {type: string, nullable: true}
        nom_projet:              {type: string, nullable: true}
        version_projet:          {type: string, nullable: true}
        branche:                 {type: string, nullable: true}
        commit_hash:             {type: string, nullable: true}
        environnement:           {type: string}
        final_statut:            {type: string, enum: [EN_ATTENTE,EN_COURS,SUCCES,ECHEC]}
        ci_statut:               {type: string, enum: [EN_ATTENTE,EN_COURS,SUCCES,ECHEC]}
        cd_statut:               {type: string, enum: [EN_ATTENTE,EN_COURS,SUCCES,ECHEC]}
        package_hash:            {type: string, nullable: true}
        nom_package:             {type: string, nullable: true}
        commence_a:              {type: string, format: date-time, nullable: true}
        fini_a:                  {type: string, format: date-time, nullable: true}
        duree:                   {type: integer, nullable: true, description: En secondes}
        created_at:              {type: string, format: date-time}
        projet:
          nullable: true
          properties:
            id:  {type: integer}
            nom: {type: string}
        declenche_par:
          nullable: true
          properties:
            id:     {type: integer}
            nom:    {type: string}
            prenom: {type: string}
        deploye_sur_serveur_par:
          nullable: true
          properties:
            id:     {type: integer}
            nom:    {type: string}
            prenom: {type: string}
    DeploiementDetail:
      allOf:
        - $ref: '#/components/schemas/DeploiementListItem'
        - type: object
          properties:
            logs:
              type: string
              description: Logs CD compilés (lu depuis logs.contenu_cd WHERE source=CD AND etape=COMPILE)
            logs_ci:
              type: string
              nullable: true
              description: Logs CI compilés (lu depuis logs.contenu_ci WHERE source=CI AND etape=COMPILE)
            declenche_par:
              type: string
              nullable: true
              description: "Nom complet ex: Emmy Admin"
            deploye_sur_serveur_par:
              type: string
              nullable: true
...